Guilds mechanic has been extracted into trainers.cpp gile from spec_procs.cpp
