+++ lib.template/text/help/feats.hlp
@@ -121,18 +121,6 @@
 +------------------------------------------------------------------------------+
 
 #
-СПОСОБНОСТЬМАГИЧЕСКОЕРОДСТВО МАГИЧЕСКОЕРОДСТВО
-+------------------------------------------------------------------------------+
-| Способность: $COLORGмагическое родство$COLORn                                              |
-+------------------------------------------------------------------------------+
-| Вы  обладаете  особым  чутьем  к  волшебным силам, что позволяет Вам творить |
-| заклинания  даже  лучше,  чем  потратившим  годы на изучение магических наук |
-| магам. (Улучшенный вариант способности "$COLORGмагическое чутье$COLORn")                   |
-+------------------------------------------------------------------------------+
-| Увеличивает шанс удачного колдовства на 80.                                  |
-+------------------------------------------------------------------------------+
-
-#
 СПОСОБНОСТЬСЛЕДОПЫТ СЛЕДОПЫТ
 +------------------------------------------------------------------------------+
 | Способность: $COLORGследопыт$COLORn                                                        |
@@ -256,17 +244,6 @@
 | Вам под горячую руку лучше не попадаться. Да и вообще тоже.                  |
 +------------------------------------------------------------------------------+
 | Увеличивает повреждение на 2.                                                |
-+------------------------------------------------------------------------------+
-
-#
-СПОСОБНОСТЬМАГИЧЕСКОЕЧУТЬЕ МАГИЧЕСКОЕЧУТЬЕ
-+------------------------------------------------------------------------------+
-| Способность: $COLORGмагическое чутье$COLORn                                                |
-+------------------------------------------------------------------------------+
-| Вы  имеете  некоторые  врожденные  способности  к магии, что позволяет лучше |
-| творить заклинания.                                                          |
-+------------------------------------------------------------------------------+
-| Увеличивает успех колдовства на 30.                                          |
 +------------------------------------------------------------------------------+
 
 #
