+++ src/spec_assign.cpp
@@ -14,8 +14,8 @@
 
 #include "obj_prototypes.h"
 #include "cmd/mercenary.h"
-#include "constants.h"
-#include "utils/logger.h"
+//#include "constants.h"
+//#include "utils/logger.h"
 #include "house.h"
 #include "boards/boards_constants.h"
 #include "boards/boards.h"
